harmonize/requirements.txt
fastapi==0.111.0
mutagen==1.47.0
yt-dlp==2024.7.1
youtube-search-python==1.6.6
Pillow==10.4.0
requests==2.32.3
ruff>=0.5.1
pytest>=8.3.3
sqlmodel>=0.0.22 
aiohttp==3.10.5
beautifulsoup4==4.11.2
eyed3==0.9.7 
paramiko>=2.11.0,<3.0.0